flask-server/requirements.txt
flask==2.3.3
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session==0.5.0
Werkzeug==2.3.8
redis
Flask-Mail